+++ src/features/Home/index.js
@@ -5,8 +5,8 @@ import Head from 'common/components/Head';
 import Container from 'common/components/Container';
 
 import styles from './styles.less';
+import AsyncComponent from 'common/components/Async';
 import Introduction from './components/Introduction';
-import RandomCharacter from './components/RandomCharacter';
 import RandomGuilds from './components/RandomGuilds';
 import ResponsiveLeaderboard from 'common/components/DisplayAd/ResponsiveLeaderboard';
 
@@ -17,7 +17,11 @@ const Home = () => (
     <div className={styles.introBackground}>
       <Container className={styles.atfContainer}>
         <Introduction className={styles.introContainer} />
-        <RandomCharacter type="ofTheDay" />
+
+        <AsyncComponent
+          type="ofTheDay"
+          load={() => import('./components/RandomCharacter')}
+        />
       </Container>
     </div>